@@ -1,5 +1,6 @@
 import React, { createContext, useState, useEffect } from 'react';
 import axios from 'axios';
+import { useNavigate } from 'react-router-dom';
 
 const AuthContext = createContext();
 
@@ -7,13 +8,14 @@ export const AuthProvider = ({ children }) => {
   const [user, setUser] = useState(null);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
+  const navigate = useNavigate();
 
   useEffect(() => {
     const checkAuth = async () => {
       try {
         const token = localStorage.getItem('token');
         if (token) {
-          const response = await axios.get('/api/users/me', {
+          const response = await axios.get('http://localhost:3000/api/users/me', {
             headers: { Authorization: `Bearer ${token}` }
           });
           setUser(response.data);
@@ -31,10 +33,11 @@ export const AuthProvider = ({ children }) => {
   const login = async (email, password) => {
     try {
       setLoading(true);
-      const response = await axios.post('/api/users/login', { email, password });
+      const response = await axios.post('http://localhost:3000/api/users/login', { email, password });
       setUser(response.data.user);
       localStorage.setItem('token', response.data.token);
       axios.defaults.headers.common['Authorization'] = `Bearer ${response.data.token}`;
+      navigate('/home');
     } catch (error) {
       setError('Login failed');
       throw new Error('Login failed');
@@ -47,6 +50,7 @@ export const AuthProvider = ({ children }) => {
     setUser(null);
     localStorage.removeItem('token');
     delete axios.defaults.headers.common['Authorization'];
+    navigate('/');
   };
 
   return (
